Feat: 함수 songInfo쪽으로 이동(노래와 관련된 함수이기떄문에)
